Allow Viewfinder data to be saved as TIFF
